# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Definition Service
  version: "1.0.0"
  description: >
    Manages workflow template lifecycle: create, draft, publish, archive, clone,
    promote, validate, export, and diff BPMN-based workflow definitions.
    All identity and tenant context is injected by the Envoy API Gateway.
servers:
  - url: http://localhost:8080
    description: Local dev server (matches cmd/server's actual route mounts)
  - url: https://api.internal/definition/v1
    description: Envoy API Gateway (production)
# All endpoints require a valid tenant context injected by the Envoy gateway.
# Mutating endpoints (POST/PUT/DELETE) additionally require x-tenant-roles.
security:
  - GatewayTenantContext: []
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: >
        Tenant UUID injected by the Envoy ext_authz filter after JWT validation.
        Must be present on every request. Never supplied directly by the client
        in production — the gateway rejects any request without a valid JWT and
        sets this header from the token claims.

    AdminRoles:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Roles claim injected by the gateway. Required for mutating operations.
        Contains a comma-separated list of roles (e.g. tenant_admin,member).
        Accepted values for mutation: tenant_admin, tenant_owner.
        Read-only endpoints (GET) do not require this header.

    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (e.g. POST /internal/events).
        These routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by Kubernetes NetworkPolicy and
        authenticated by this shared secret (or mTLS, depending on deployment). The tenant context
        for these routes is taken from the request payload, not from x-tenant-id.
  parameters:
    TenantId:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID injected by Envoy.
    UserId:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID injected by Envoy.
    TenantRoles:
      name: x-tenant-roles
      in: header
      required: true
      schema:
        type: string
      description: Comma-separated tenant roles injected by Envoy.
    Departments:
      name: x-departments
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated list of department memberships and role levels injected by Envoy.
    Plan:
      name: x-plan
      in: header
      required: false
      schema:
        type: string
      description: Tenant billing plan tier injected by Envoy.
    FeatureFlags:
      name: x-feature-flags
      in: header
      required: false
      schema:
        type: string
      description: Encoded set of effective feature flags injected by Envoy.
    WorkflowId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional unique client-generated request key to prevent duplicate operations on network retries.
  schemas:
    VersionStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
...
    Pagination:
      type: object
      properties:
        total_count: { type: integer }
        page:        { type: integer }
        limit:       { type: integer }
      example:
        total_count: 4
        page: 1
        limit: 20
    ValidationError:
      type: object
      properties:
        node_id: { type: string }
        error:   { type: string }
      example:
        node_id: Task_design_prep
        error: "userTask is missing zeebe:taskDefinition"
...
    CatalogScope:
      type: string
      enum: [tenant, global]
      description: >-
        `tenant` rows are owned by the caller's own tenant; `global` rows are
        platform-authored and visible to every tenant. Creating or deleting a
        `global` row requires the `platform_operator` role.
    ModuleSummary:
      type: object
      properties:
        id:                { type: string, format: uuid }
        tenant_id:         { type: string, format: uuid, nullable: true }
        scope:             { $ref: '#/components/schemas/CatalogScope' }
        name:              { type: string }
        description:       { type: string }
        active_version_id: { type: string, format: uuid, nullable: true }
        created_at:        { type: string, format: date-time }
        updated_at:        { type: string, format: date-time }
      example:
        id: 018f1a2b-0000-7000-8000-000000000001
        tenant_id: null
        scope: global
        name: Standard Approval Chain
        description: Reusable three-step approval sequence
        active_version_id: 018f1a2b-0000-7000-8000-000000000002
        created_at: "2026-08-20T08:00:00Z"
        updated_at: "2026-08-25T10:00:00Z"
    ModuleVersion:
      type: object
      properties:
        id:              { type: string, format: uuid }
        module_id:       { type: string, format: uuid }
        scope:           { $ref: '#/components/schemas/CatalogScope' }
        status:          { $ref: '#/components/schemas/VersionStatus' }
        bpmn_xml:        { type: string }
        process_id:
          type: string
          description: The `<bpmn:process id="...">` value inside `bpmn_xml`, matched against a referencing diagram's `callActivity` `calledElement`.
        version_number:  { type: integer, nullable: true }
        is_valid:        { type: boolean }
        published_at:    { type: string, format: date-time, nullable: true }
        created_at:      { type: string, format: date-time }
      example:
        id: 018f1a2b-0000-7000-8000-000000000002
        module_id: 018f1a2b-0000-7000-8000-000000000001
        scope: global
        status: PUBLISHED
        process_id: ApprovalChain
        version_number: 1
        is_valid: true
        published_at: "2026-08-20T09:00:00Z"
        created_at: "2026-08-20T08:00:00Z"
    ModuleDetail:
      allOf:
        - $ref: '#/components/schemas/ModuleSummary'
        - type: object
          properties:
            active_version:
              allOf: [{ $ref: '#/components/schemas/ModuleVersion' }]
              nullable: true
...
    CompiledPlan:
      type: object
      required: [name, departments, execution]
      properties:
        name:        { type: string }
        task_queue:  { type: string, description: "Temporal task queue. Omitted for ignored pools." }
        ignored:
          type: boolean
          description: "True when this pool was marked ignore=true in the BPMN. Consumers decide how to handle it."
        departments:
          type: array
          items: { $ref: '#/components/schemas/DepartmentDef' }
        execution:
          type: object
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/ExecutionStep' }
        visual_elements:
          type: array
          nullable: true
          description: "Visual-only BPMN elements (e.g. dataStoreReference) that carry no execution semantics. Informational only."
          items: { $ref: '#/components/schemas/VisualElementDef' }
      example:
        name: Tender Document Review Process
        task_queue: wf-queue-default
        departments:
          - id: Tender
            label: Tender
            stages:
              - { type: prep,    activity: PrepActivity,    role: bd-agent,       default_assignees: [018e1f2a-0000-7000-8000-000000000001], extras: { requires_comment: "false" } }
              - { type: review,  activity: ReviewActivity,  role: tender-manager, default_assignees: [018e1f2a-0000-7000-8000-000000000002], extras: { requires_comment: "true" } }
              - { type: approve, activity: ApproveActivity, role: tender-director, default_assignees: [018e1f2a-0000-7000-8000-000000000003], extras: { requires_comment: "true" } }
          - id: Engineering
            label: Engineering
            stages:
              - { type: prep,    activity: PrepActivity,    role: lead-engineer,  default_assignees: [018e1f2a-0000-7000-8000-000000000010], extras: { requires_comment: "false" }, boundary_timer: { duration: "PT48H", interrupting: false } }
        execution:
          steps:
            - sequential: [Tender, Engineering]
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem.
        code:
          type: string
          description: Service-specific error code for programmatic handling.
        invalid_params:
          type: array
          description: Details about specific fields that failed validation.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
              reason:
                type: string
              code:
                type: string
      example:
        type: "https://api.workflow.platform/errors/not-found"
        title: "Resource Not Found"
        status: 404
        detail: "Workflow not found or tenant boundary breached."
        instance: "/workflows/8fa88cde-824c-47bc-836b-665cd42c2222"
        code: "NOT_FOUND"
...
  responses:
    Unauthorized:
      description: Missing or invalid gateway identity headers.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/unauthorized"
            title: "Unauthorized Access"
            status: 401
            detail: "x-tenant-id header is missing or could not be resolved to a valid tenant."
            instance: "/workflows"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Caller lacks the required role or plan limits were exceeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/forbidden"
            title: "Forbidden Operation"
            status: 403
            detail: "This operation requires x-tenant-roles: tenant_admin or tenant_owner."
            instance: "/workflows"
            code: "FORBIDDEN"
    NotFound:
      description: Resource not found or tenant boundary breached (security through obscurity).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/not-found"
            title: "Resource Not Found"
            status: 404
            detail: "Workflow not found or belongs to a different tenant."
            instance: "/workflows/8fa88cde-824c-47bc-836b-665cd42c2222"
            code: "NOT_FOUND"
...
    UnprocessableEntity:
      description: Semantically invalid BPMN — see errors array for per-node details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/validation-failed"
            title: "BPMN Validation Failed"
            status: 422
            detail: "BPMN semantic or structural validation failed."
            instance: "/workflows/8fa88cde-824c-47bc-836b-665cd42c2222/versions/11abcc22/publish"
            code: "BPMN_VALIDATION_FAILED"
            invalid_params:
              - name: Task_design_prep
                reason: "userTask is missing zeebe:taskDefinition"
                code: MISSING_TASK_DEFINITION
              - name: Gateway_parallel_split
                reason: "Parallel split gateway has no reachable matching join gateway"
                code: UNMATCHED_GATEWAY
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred. Please retry or contact support if the issue persists."
            instance: "/workflows"
            code: "INTERNAL_ERROR"
paths:
...
  /api/v1/workflows/{id}/versions/{version_id}:
    get:
      operationId: getVersion
      summary: Get version detail including BPMN XML and compiled DSL
      tags: [Versions]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/Departments'
        - $ref: '#/components/parameters/Plan'
        - $ref: '#/components/parameters/FeatureFlags'
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        "200":
          description: Version detail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  version_id:         { type: string, format: uuid }
                  workflow_id:        { type: string, format: uuid }
                  version_number:     { type: integer, nullable: true }
                  status:             { $ref: '#/components/schemas/VersionStatus' }
                  is_valid:           { type: boolean }
                  validation_errors:
                    type: array
                    items: { $ref: '#/components/schemas/ValidationError' }
                    nullable: true
                  bpmn_xml:           { type: string }
                  module_bpmn_xmls:   { type: array, items: { type: string }, description: "Called-process BPMNs attached to this version. Empty array if none." }
                  artifact_hash:      { type: string, nullable: true, description: "SHA-256 of the BPMN XML at publish time. Absent on DRAFT versions." }
                  compiled_plan_json: { $ref: '#/components/schemas/CompiledPlan' }
                  created_by_user_id: { type: string, format: uuid }
                  created_at:         { type: string, format: date-time }
                  published_at:       { type: string, format: date-time, nullable: true }
                  record_version:     { type: integer, format: int64, description: "Monotonic counter. Echo on PUT /draft requests." }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalError' }

  # 3.7
...
  /api/v1/modules:
    get:
      operationId: listModules
      summary: List modules visible to the caller (own tenant + global)
      tags: [Modules]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: scope
          schema: { $ref: '#/components/schemas/CatalogScope' }
        - in: query
          name: q
          schema: { type: string }
          description: Fuzzy search on name.
      responses:
        "200":
          description: Paginated module list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  modules:
                    type: array
                    items: { $ref: '#/components/schemas/ModuleSummary' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        "401": { $ref: '#/components/responses/Unauthorized' }

    post:
      operationId: createModule
      summary: Create a module and its first DRAFT version
      description: >
        `scope: global` requires the `platform_operator` role and is written
        through a BYPASSRLS system connection — the ordinary tenant-scoped
        role cannot insert a global-scope row regardless of this check.
      tags: [Modules]
      security:
        - GatewayTenantContext: []
          AdminRoles: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, bpmn_xml]
              properties:
                name:        { type: string, maxLength: 255 }
                description: { type: string, maxLength: 2000 }
                bpmn_xml:
                  type: string
                  description: Must contain exactly one top-level `<bpmn:process>`.
                scope:
                  allOf: [{ $ref: '#/components/schemas/CatalogScope' }]
                  default: tenant
      responses:
        "201":
          description: Module and initial draft version created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  module_id:  { type: string, format: uuid }
                  version_id: { type: string, format: uuid }
                  status:     { $ref: '#/components/schemas/VersionStatus' }
                  message:    { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalError' }
  /api/v1/modules/{id}:
    get:
      operationId: getModule
      summary: Get a module's metadata and active version
      tags: [Modules]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Module detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModuleDetail' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: archiveModule
      summary: Archive a module (clears its active version)
      description: Deleting a `scope=global` module requires the `platform_operator` role.
      tags: [Modules]
      security:
        - GatewayTenantContext: []
          AdminRoles: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Module archived.
          content:
            application/json:
              schema:
                type: object
                properties:
                  module_id: { type: string, format: uuid }
                  status:    { type: string, example: ARCHIVED }
                  message:   { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409":
          description: The module has no active version to archive.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/v1/modules/{id}/versions:
    get:
      operationId: listModuleVersions
      summary: List a module's version history
      tags: [Modules]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated version list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items: { $ref: '#/components/schemas/ModuleVersion' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        "401": { $ref: '#/components/responses/Unauthorized' }

    post:
      operationId: addModuleVersion
      summary: Add a new DRAFT version to an existing module
      description: >
        `bpmn_xml` is optional — omit it to copy the current active version's
        content forward, or supply it to author fresh content in the new draft.
      tags: [Modules]
      security:
        - GatewayTenantContext: []
          AdminRoles: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                bpmn_xml: { type: string, nullable: true }
      responses:
        "201":
          description: Draft version created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  module_id:  { type: string, format: uuid }
                  version_id: { type: string, format: uuid }
                  status:     { $ref: '#/components/schemas/VersionStatus' }
                  message:    { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409":
          description: A DRAFT version already exists for this module.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/modules/{id}/versions/{version_id}:
    get:
      operationId: getModuleVersion
      summary: Get a specific module version, including its BPMN XML
      tags: [Modules]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: version_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Module version detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModuleVersion' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/v1/modules/{id}/versions/{version_id}/publish:
    post:
      operationId: publishModuleVersion
      summary: Publish a DRAFT module version
      tags: [Modules]
      security:
        - GatewayTenantContext: []
          AdminRoles: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: version_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Version published and set as the module's active version.
          content:
            application/json:
              schema:
                type: object
                properties:
                  module_id:      { type: string, format: uuid }
                  version_id:     { type: string, format: uuid }
                  status:         { $ref: '#/components/schemas/VersionStatus' }
                  version_number: { type: integer }
                  message:        { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409":
          description: The version is not in DRAFT status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
...